召唤百分比增伤可以应用于召唤物；联动铁魔法，召唤物造成额外伤害；联动curiosAPI，添加坐骑和宠物饰品栏和快捷键；添加召唤师头盔附魔，和保护冲突

+++ src/main/java/org/confluence/terraentity/entity/proj/WhipEntity.java
@@ -228,11 +228,12 @@ public class WhipEntity extends Projectile {
             }
         }
 
-        if(existTick * 0.3f == tickCount){
+        if((int)(existTick * 0.3f) == tickCount){
             if(this.getOwner() != null) {
                 this.getOwner().playSound(TESounds.WHIP_ATTACK.get(), 0.6F + getRandom().nextFloat() * 0.2f, 1.0F);
             }
         }
+
 
         if(parts == null || parts.isEmpty()) return;
         this.speed = (double) _existTick / this.existTick;
